minor, provide converter of echse syntax -> state syntax
The working title of the tool at the moment is echse-state
but functionality will be joined into the main (and installed)
echse utility.

Like normal echse events, states will only appear in the output
as edges, i.e. when the state actually changes.

At the moment there is no syntactic way to distinguish echse and
state files.

Signed-off-by: Sebastian Freundt <freundt@fresse.org>