hygiene, do away with the ECHS_CMD_MORE concept as it's faulty ...
It doesn't store which of the tried command predicates has thrown it
and hence the next time data arrives we have no choice but
to try them all again.  However, some of the routines might
have lodged some data in the param context according to their
previous findings (to keep some state) but this will prove
problematic (it's a union) when several other predicates
inspect these values cast to their reckoning of the param
closure.  In short, it's a completely idiotic thing.