minor, to-be-reverted, provide rdate and exdate setter ...
This whole endeavour is illogical though.  We have
decided to make the c2s communication CRUD and stateless,
so updates of just one property (though dictated by
RFC 5546) are just stupid.

The new way of cancellation of single recurrences will be
to retrieve the task in question, then manipulate it on
the client side (with no immediate help from libechse) and
send it back to the server as an update.

Expect reversions in the next few patches.